backend user validation

+++ app/api/restaurant_routes.py
@@ -125,8 +125,8 @@ def delete_item(item_id):
     if not menu_item:
         return {"message": "Menu Item couldn't be found"}
 
-    # if int(current_user.get_id()) != menu_item.restaurant.user_id:
-    #     return {"errors": {"message": "Unauthorized"}}, 401
+    if int(current_user.get_id()) != menu_item.restaurant.user_id:
+        return {"errors": {"message": "Unauthorized"}}, 401
 
     db.session.delete(menu_item)
     db.session.commit()
